+++ app/routes/components/Button/changelog.codegen.ts
@@ -17,4 +17,6 @@ export const changelog = ''
   + '- 2025-10-15: Support change styles when `disabled`.\n'
   + '- 2025-10-20: Support the prop `color`.\n'
   + '- 2025-10-28: Add `secondary` variant.\n'
+  + '- 2025-10-30: Add `--button-border-separator-color` CSS variable and add the\n'
+  + '  default icon size for SVG children in left or right section.\n'
   ;