+++ src/components/Button/CHANGELOG.md
@@ -13,3 +13,5 @@
 - 2025-10-15: Support change styles when `disabled`.
 - 2025-10-20: Support the prop `color`.
 - 2025-10-28: Add `secondary` variant.
+- 2025-10-30: Add `--button-border-separator-color` CSS variable and add the
+  default icon size for SVG children in left or right section.